move user affiliation mutation tests
